refactor out geolookup, fix spacing listing
